+++ packages/orq-rc/src/orq_ai_sdk/contacts.py
@@ -3,10 +3,6 @@
 from .basesdk import BaseSDK
 from orq_ai_sdk import models, utils
 from orq_ai_sdk._hooks import HookContext
-from orq_ai_sdk.models import (
-    createcontactop as models_createcontactop,
-    listcontactsop as models_listcontactsop,
-)
 from orq_ai_sdk.types import BaseModel, OptionalNullable, UNSET
 from orq_ai_sdk.utils import get_security_from_env
 from orq_ai_sdk.utils.unmarshal_json_response import unmarshal_json_response
@@ -19,8 +15,8 @@ class Contacts(BaseSDK):
         *,
         request: Optional[
             Union[
-                models_createcontactop.CreateContactRequestBody,
-                models_createcontactop.CreateContactRequestBodyTypedDict,
+                models.CreateContactRequestBody,
+                models.CreateContactRequestBodyTypedDict,
             ]
         ] = None,
         retries: OptionalNullable[utils.RetryConfig] = UNSET,
@@ -116,8 +112,8 @@ class Contacts(BaseSDK):
         *,
         request: Optional[
             Union[
-                models_createcontactop.CreateContactRequestBody,
-                models_createcontactop.CreateContactRequestBodyTypedDict,
+                models.CreateContactRequestBody,
+                models.CreateContactRequestBodyTypedDict,
             ]
         ] = None,
         retries: OptionalNullable[utils.RetryConfig] = UNSET,
@@ -216,10 +212,7 @@ class Contacts(BaseSDK):
         ending_before: Optional[str] = None,
         search: Optional[str] = None,
         filter_by: Optional[
-            Union[
-                models_listcontactsop.QueryParamFilterBy,
-                models_listcontactsop.QueryParamFilterByTypedDict,
-            ]
+            Union[models.QueryParamFilterBy, models.QueryParamFilterByTypedDict]
         ] = None,
         include_metrics: OptionalNullable[bool] = False,
         retries: OptionalNullable[utils.RetryConfig] = UNSET,
@@ -325,10 +318,7 @@ class Contacts(BaseSDK):
         ending_before: Optional[str] = None,
         search: Optional[str] = None,
         filter_by: Optional[
-            Union[
-                models_listcontactsop.QueryParamFilterBy,
-                models_listcontactsop.QueryParamFilterByTypedDict,
-            ]
+            Union[models.QueryParamFilterBy, models.QueryParamFilterByTypedDict]
         ] = None,
         include_metrics: OptionalNullable[bool] = False,
         retries: OptionalNullable[utils.RetryConfig] = UNSET,