+++ packages/orq-rc/src/orq_ai_sdk/models/conversationresponse.py
@@ -70,7 +70,7 @@ class ConversationResponse(BaseModel):
     r"""Unix timestamp (in milliseconds) when the conversation was last updated"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "conv_01kdyvszkw27tg5fs5zerxyywh"
+        "conv_01kdyzdpq9j2skkx1tnv0k65s3"
     )
     r"""Unique ULID (Universally Unique Lexicographically Sortable Identifier) for the conversation, prefixed with \"conv_\" """