+++ packages/orq-rc/src/orq_ai_sdk/models/createconversationop.py
@@ -104,7 +104,7 @@ class CreateConversationResponseBody(BaseModel):
     r"""Unix timestamp (in milliseconds) when the conversation was last updated"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "conv_01kdyvszkzp3et86kd0b1zm7rn"
+        "conv_01kdyzdpqbd0vhg4971003vp1y"
     )
     r"""Unique ULID (Universally Unique Lexicographically Sortable Identifier) for the conversation, prefixed with \"conv_\" """