+++ packages/orq-rc/src/orq_ai_sdk/models/retrievedatasourceop.py
@@ -90,7 +90,7 @@ class RetrieveDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KDYVT018F0AN8ZV63EF6RDD7"
+        "01KDYZDQ5NXWEHGE7SVXR26G9B"
     )
     r"""The unique identifier of the data source"""