+++ packages/orq-rc/RELEASES.md
@@ -4519,3 +4519,13 @@ Based on:
 - [python v4.2.0-rc.7] packages/orq-rc
 ### Releases
 - [PyPI v4.2.0-rc.7] https://pypi.org/project/orq-ai-sdk/4.2.0-rc.7 - packages/orq-rc
+
+## 2026-01-02 09:10:33
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.680.11 (2.788.15) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v4.2.0-rc.8] packages/orq-rc
+### Releases
+- [PyPI v4.2.0-rc.8] https://pypi.org/project/orq-ai-sdk/4.2.0-rc.8 - packages/orq-rc